+++ spring-data-elasticsearch-rest/src/test/java/com/github/ydespreaux/spring/data/elasticsearch/core/indices/IndexBuilderTest.java
@@ -20,26 +20,39 @@
 
 package com.github.ydespreaux.spring.data.elasticsearch.core.indices;
 
-import org.elasticsearch.action.admin.indices.alias.Alias;
 import org.elasticsearch.action.admin.indices.create.CreateIndexRequest;
-import org.elasticsearch.common.settings.Settings;
 import org.junit.Test;
 import org.springframework.core.io.ClassPathResource;
 
-import java.util.Map;
-import java.util.Set;
+import java.util.ArrayList;
+import java.util.Collections;
+
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertThat;
 
 public class IndexBuilderTest {
 
     @Test
-    public void shouldFailToInitializeGivenMappingContextIsNull() {
+    public void createIndexRequestToInitializeGivenResource() {
         // given
-        CreateIndexBuilder builder = new CreateIndexBuilder().name("books").source(new ClassPathResource("indices/book.index"));
+        CreateIndexBuilder builder = new CreateIndexBuilder().name("music").source(new ClassPathResource("indices/music.index"));
         CreateIndexRequest request = builder.build();
-        Set<Alias> aliases = request.aliases();
-        Settings settings = request.settings();
-        Map<String, String> mappings = request.mappings();
-        System.out.println(mappings);
+        assertThat(new ArrayList<>(request.aliases()).get(0).name(), is(equalTo("musics")));
+        assertThat(request.settings().get("number_of_shards"), is(equalTo("1")));
+        assertThat(request.settings().get("refresh_interval"), is(equalTo("1s")));
+        assertThat(request.settings().get("number_of_replicas"), is(equalTo("1")));
+        assertThat(request.settings().get("store.type"), is(equalTo("fs")));
+        assertThat(request.mappings().containsKey("album"), is(true));
     }
 
+    @Test
+    public void createIndexRequestToInitializeGivenNoResource() {
+        // given
+        CreateIndexBuilder builder = new CreateIndexBuilder().name("index_name").sources(Collections.emptyList());
+        CreateIndexRequest request = builder.build();
+        assertThat(request.aliases().isEmpty(), is(true));
+        assertThat(request.settings().isEmpty(), is(true));
+        assertThat(request.mappings().isEmpty(), is(true));
+    }
 }